+++ src/main/java/com/github/kd_gaming1/skyblockenhancements/config/SkyblockEnhancementsConfig.java
@@ -19,6 +19,9 @@ public class SkyblockEnhancementsConfig extends MidnightConfig {
     public static boolean enterToConfirmSign = true;
 
     @Entry(category = SKYBLOCK_ENHANCEMENTS)
+    public static boolean enterToConfirmAllSigns = false;
+
+    @Entry(category = SKYBLOCK_ENHANCEMENTS)
     public static boolean enableItemGlowOutline = true;
 
     @Entry(category = SKYBLOCK_ENHANCEMENTS)